paralel stream, measure helper, faster loading

+++ src/main/java/lu/crx/financing/services/BatchFinancingService.java
@@ -3,5 +3,5 @@ package lu.crx.financing.services;
 import lu.crx.financing.util.InvoiceFactoringProcess;
 
 public interface BatchFinancingService {
-    int financeBatch(InvoiceFactoringProcess invoiceFactoringProcess);
+    long financeBatch(InvoiceFactoringProcess invoiceFactoringProcess);
 }
